crawler_comercial_escritorio_vendas/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

boto3==1.26.92
office365-rest-client==2.2.0
pandas==1.5.1
SharePlum==0.5.1
openpyxl==3.1.2
pyarrow==11.0.0